Code_Auto_Generation_Agent/bak/pyproject.toml
[project]
name = "auto-coding-agent"
version = "0.1.0"
description = "Automatic Code Generation Agent - generate full project from PRD json using AI"
requires-python = ">=3.10"
dependencies = [
  "python-dotenv>=1.0.0",
  "fastapi>=0.124.4",
  "uvicorn[standard]>=0.33.0",
  "python-multipart>=0.0.20",
  "jinja2>=3.1.0",
  "pydantic>=2.0.0",
  "openai>=1.0.0",
  "langchain>=0.3.0",
  "langchain-openai>=0.2.0",
  "langgraph>=0.2.0",
]

[project.scripts]
auto-coding = "auto_coding:main"

[tool.ruff]
line-length = 100
select = ["E", "W", "F", "I"]
ignore = ["E501"]

# output 目录下是自动生成的独立项目，不加入工作区
# 避免依赖污染和路径混淆
[tool.uv.workspace]
members = []
exclude = ["output/*"]

[dependency-groups]
dev = [
    "mypy>=1.20.1",
    "pytest>=9.0.3",
    "ruff>=0.15.11",
]
